converted get_reports() to SOAP interface
Now we return debianbts.Bugreport in the hierarchy, so the UI can extract the
information it needs and display them, instead of "forging" a string the UIs
need to parse.

Converting to SOAP and object attributes for gathering data, simplified a lot
the code for the UIs, opening the possibility to drop a lot of dead code in the
debbugs module (SMGL parsing, for example).