when using --src don't present the menu with possible binary packages (and also try to be smart and infer the version from the bin pkgs list); thanks to Vincent Lefevre for the report; Closes: #638616
In case --src is used, we don't want to suggests bin pkgs, but we translate the
pkg passed as args into its source pkg (already done) and then we short-circuit
the checks.

The trick for getting the src pkg version right (since we don't have a utils
function to retrieve source packages information, like for bin) is done looping
over the bin pkgs (that we retrive anyhow) and find the first version, and use
it as the src one.