make get_report() return the Bugreport instance, containing all the bug information; now the UI does the visualization part, and we just pass objects around storing the info the UI will extract and display; implicitly Closes: #571983
This patch made several changes, removing really old/ugly code. Passing
Bugreport around allows us to query that object and not the subject/bodies of
the but report for information about it.

Something is really innatural, like keeping 'exinfo' variable around: we should
probably rename it at some point in the future, but we'd preferred taking the
safe side and not remave that too.