remove the description from the dependencies list, leaving more space for the useful information (package name and version); thanks to Andreas Beckmann for the report; Closes: #582218
The trick here is to dynamically calculate the width for the package and version
columns, in relation with the other one.

We want to store as much info as possible, but we don't want to use a line
longer than 80 chars, and also without a strict width.

So we compute the column width as the minimum between the max length of the
actual values to show or the space left by the other column.